Adds Application plugin structure
