+++ bundles/pixi.js/package.json
@@ -49,7 +49,6 @@
     "@pixi/loaders": "^5.0.0-alpha.2",
     "@pixi/math": "^5.0.0-alpha.2",
     "@pixi/mesh": "^5.0.0-alpha.2",
-    "@pixi/mixin-app-loader": "^5.0.0-alpha.2",
     "@pixi/mixin-cache-as-bitmap": "^5.0.0-alpha.2",
     "@pixi/mixin-get-child-by-name": "^5.0.0-alpha.2",
     "@pixi/mixin-get-global-position": "^5.0.0-alpha.2",
