+++ packages/loaders/README.md
@@ -8,6 +8,25 @@ npm install @pixi/loaders
 
 ## Usage
 
+Using the loader directly:
+
 ```js
-import '@pixi/loaders';
+import { Loader } from '@pixi/loaders';
+
+const loader = new Loader();
+loader.add('path/to/file.jpg');
+loader.load(() => {});
+```
+
+Using the loader as an Application plugin:
+
+```js 
+import { LoaderPlugin } from '@pixi/loaders';
+import { Application } from '@pixi/app';
+
+Application.registerPlugin(LoaderPlugin);
+
+const app = new Application();
+app.loader.add('path/to/file.jpg');
+app.loader.load(() => {});
 ```
