+++ packages/loaders/src/index.js
@@ -9,3 +9,4 @@ import { Resource } from 'resource-loader';
 export const LoaderResource = Resource;
 export { default as Loader } from './Loader';
 export { default as TextureLoader } from './TextureLoader';
+export { default as LoaderPlugin } from './LoaderPlugin';
