+++ packages/loaders/test/index.js
@@ -1,2 +1,3 @@
 require('./Loader');
 require('./TextureLoader');
+require('./LoaderPlugin');
