--- packages/mixin-app-loader/README.md
@@ -1,13 +1,0 @@
-# @pixi/mixin-app-loader
-
-## Installation
-
-```bash
-npm install @pixi/mixin-app-loader
-```
-
-## Usage
-
-```js
-import '@pixi/mixin-app-loader';
-```
