--- packages/mixin-app-loader/package.json
@@ -1,36 +1,0 @@
-{
-  "name": "@pixi/mixin-app-loader",
-  "version": "5.0.0-alpha.2",
-  "main": "lib/mixin-app-loader.js",
-  "module": "lib/mixin-app-loader.es.js",
-  "description": "Support for loader in Application",
-  "author": "Mat Groves",
-  "contributors": [
-    "Matt Karl <matt@mattkarl.com>"
-  ],
-  "homepage": "http://pixijs.com/",
-  "bugs": "https://github.com/pixijs/pixi.js/issues",
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/pixijs/pixi.js.git"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "scripts": {
-    "test": "floss --path test"
-  },
-  "files": [
-    "lib"
-  ],
-  "dependencies": {
-    "@pixi/app": "^5.0.0-alpha.2",
-    "@pixi/loaders": "^5.0.0-alpha.2"
-  },
-  "devDependencies": {
-    "@pixi/canvas-renderer": "^5.0.0-alpha.2",
-    "@pixi/utils": "^5.0.0-alpha.2",
-    "floss": "^2.1.3"
-  }
-}
