+++ packages/ticker/src/index.js
@@ -1,4 +1,5 @@
 export { default as Ticker } from './Ticker';
+export { default as TickerPlugin } from './TickerPlugin';
 export * from './const';
 
 import './settings';
